Substrate Updates (#468)
* Regenesis Pangolin (#459)

* Substrate Updates (#454)

* companion: for #substrate-7250

* lock: deps

* bump: smallvec

* update: format

* companion: for #substrate-7228

* companion: for #substrate-7350

* companion: for #substrate-6869

* bump: log

* companion: for #substrate-7000

* companion: for #substrate-7296

* companion: for #substrate-7299

* companion: for #substrate-7318

* companion: for #substrate-7384

* companion: for #substrate-7394

* companion: for #substrate-7225

* companion: for #substrate-7328

* companion: for #substrate-7233

* companion: for #substrate-7381

* fix: compile with latest rust

* companion: for #substrate-7368

* free: deps lock

* update: deps

* support: cross compile

* regenesis

* checkout: alpha version

* regenesis

* Merge commit '1d167fd97183fca44c4c2032aee886f5a17c39f9' into pangolin

* bump: version

* companion: for #substrate-7443

* companion: for #substrate-7159

* stash test env

* stash test env

* companion: for #substrate-6912

* companion: for #substrate-7484

* companion: for #substrate-7145

* companion: for #substrate-7497

* companion: for #substrate-7186

* cargo fmt

* stash test env

* companion: for #substrate-7534

* companion: for #substrate-7380

* stash test env

* companion: for #substrate-7635

* fix address but not sure if this work

* stash test env

* companion: for #substrate-7532

* companion: for #substrate-7546

* companion: for #substrate-7599

* companion: for #substrate-7631

* remove unused file

* companion: for #substrate-7542

* stash test env

* companion: for #substrate-7152

* companion: for #substrate-7646

* stash test env

* companion: for #substrate-6629 part 1

* companion: for #substrate-6629 part 2

* companion: for #substrate-7391

* companion: for #substrate-7565

* companion: for #substrate-7652

* stash test env

* companion: for #substrate-7628

* companion: for #substrate-7710

* companion: for #substrate-7718

* companion: for #substrate-7725

* companion: for #substrate-7749

* stash test env

* companion: for #substrate-7719 part 1

* companion: for #substrate-7719 part 2

* rename

* companion: for #substrate-7789

* companion: for #substrate-7784

* companion: for #substrate-7829

* companion: for #substrate-7700

* stash test env

* stash test env

* format

* companion: for #substrate-7843

* companion: for #substrate-7866

* companion: for #substrate-7878

* companion: for #substrate-7692

* companion: for #substrate-7886

* fix: whitespace

* companion: for #substrate-7931

* update cargo spec

* fix: compile for latest rustc

* lock deps

* companion: for #substrate-7775

* companion: for #substrate-7810

* fix typo

* companion: for #substrate-7873

* companion: for #substrate-7363 part 1

* companion: for #substrate-7363 part  2

* companion: for #substrate-7363 part 3

* companion: for #substrate-7908

* companion: for #substrate-7898

* fix: staking test

* companion: for #substrate-7363 part 4

* companion: for #substrate-7463

* companion: for #substrate-7930

* simplify impl account data

* companion: for #substrate-7040

* companion: for #substrate-7933

* fix type

* pick from https://github.com/AsceticBear/darwinia-common/commit/608a6bc4715bf1d1ca26f657a99a18bba20e6c5d

* companion: for #substrate-7711

* companion: for #substrate-7339

* companion: for #substrate-7953

* cargo fmt

* fix type

* update deps

* Update ethereum-types in dvm

* remove unused import

* fix runtime

* update deps

* fix compile

* Format and fix dvm tests

* Remove duplicate files

* release CI rustc version

* update deps

* fix compile

* cargo fmt

* try fix ci

* fix ci

* migrate balances mock to construct runtime

* migrate vesting mock to construct runtime

* migrate staking mock to construct runtime

* fix old tests

* try fix ci

* migrate header mmr mock to construct runtime

* fix path

* cargo b

* migrate claims mock to construct runtime

* update ci

* migrate relay authorities mock to construct runtime

* revert

* fix typo

* fix multi address usage

* use multi address

* release deps lock

* fix kton precompile

+++ frame/evm/precompile/contracts/empty/src/lib.rs
@@ -27,10 +27,10 @@ use evm::{ExitError, ExitSucceed};
 pub struct Empty;
 
 impl LinearCostPrecompile for Empty {
-	const BASE: usize = 0;
-	const WORD: usize = 0;
+	const BASE: u64 = 0;
+	const WORD: u64 = 0;
 
-	fn execute(_: &[u8], _: usize) -> core::result::Result<(ExitSucceed, Vec<u8>), ExitError> {
+	fn execute(_: &[u8], _: u64) -> core::result::Result<(ExitSucceed, Vec<u8>), ExitError> {
 		Err(ExitError::Other("Not implement yet".into()))
 	}
 }
